stacks/titanic_model/02_deploy/requirements.txt
scikit-learn==1.5.2 
db-dtypes==1.3.0 
pandas==2.2.3 
fastapi==0.115.0
uvicorn==0.31.0
google-cloud-storage==2.18.2
